requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
chromadb>=0.5.0
neo4j>=5.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
